refs #850 Convert some unit tests to integration test
