+++ package.json
@@ -110,6 +110,11 @@
     "testMatch": [
       "**/spec/**/*.spec.ts"
     ],
+    "preset": "ts-jest/presets/js-with-ts",
+    "transform": {
+      "^.+\\.(js|jsx)$": "babel-jest",
+      "^.+\\.(ts|tsx)$": "ts-jest"
+    },
     "globals": {
       "ts-jest": {
         "tsConfig": "tsconfig.json"
