+++ spec/renderer/unit/store/Login.spec.ts
@@ -1,9 +1,4 @@
 import Login from '@/store/Login'
-import axios from 'axios'
-import { ipcMain } from '~/spec/mock/electron'
-
-jest.mock('axios')
-axios.get = jest.fn()
 
 describe('Login', () => {
   describe('mutations', () => {
@@ -28,55 +23,4 @@ describe('Login', () => {
       })
     })
   })
-
-  // TODO: move to integration
-  describe.skip('actions', () => {
-    describe('fetchLogin', () => {
-      describe('error', () => {
-        it('should return error', async () => {
-          ipcMain.once('get-auth-url', (event, _) => {
-            event.sender.send('error-get-auth-url', new AuthError())
-          })
-          const commitMock = jest.fn()
-          await Login.actions.fetchLogin({ commit: commitMock }, 'pleroma.io')
-            .catch((err) => {
-              expect(err instanceof AuthError).toEqual(true)
-            })
-        })
-      })
-      describe('success', () => {
-        it('should return url', async () => {
-          ipcMain.once('get-auth-url', (event, _) => {
-            event.sender.send('response-get-auth-url', 'http://example.com/auth')
-          })
-          const commitMock = jest.fn()
-          const url = await Login.actions.fetchLogin({ commit: commitMock }, 'pleroma.io')
-          expect(url).toEqual('http://example.com/auth')
-        })
-      })
-    })
-    describe('pageBack', () => {
-      it('should reset instance', () => {
-        const commitMock = jest.fn()
-        Login.actions.pageBack({ commit: commitMock })
-        expect(commitMock).toHaveBeenCalledWith('changeInstance', null)
-      })
-    })
-    describe('confirmInstance', () => {
-      it('should change instance', async () => {
-        // Provide Promise.resolve for finally keywrod.
-        // https://github.com/facebook/jest/issues/6552
-        (<jest.Mock>axios.get).mockReturnValue(Promise.resolve({
-          data: 'test'
-        }))
-        const commitMock = jest.fn()
-        const data = await Login.actions.confirmInstance({ commit: commitMock }, 'pleroma.io')
-        expect(data).toEqual('test')
-        // ref: https://eddyerburgh.me/how-to-unit-test-a-vuex-store
-        expect(commitMock).toHaveBeenCalledWith('changeInstance', 'pleroma.io')
-      })
-    })
-  })
 })
-
-class AuthError extends Error { }
